Grabación de vídeo en disco
Se modifica la grabación de vídeo en disco para grabar las videotecas dentro de una subcarpeta correspondiente al id del usuario. De este modo están todas agrupadas

+++ src/main/java/com/oscar/videoteca/rest/util/FileUtil.java
@@ -130,10 +130,18 @@ public class FileUtil {
 					}
 					
 				});
+				
+				exito =Boolean.TRUE;
 			} else  {
-				f.delete();
 				exito = Boolean.TRUE;
 			}
+		
+			// Un vez borrados los archivos dentro del directory, se borra el directorio
+			if(Boolean.TRUE.equals(exito)) {
+				f.delete();
+			}
+			
+			
 			
 		}
 		
@@ -194,57 +202,57 @@ public class FileUtil {
 
 	
 
-	/**
-	 * Devuelve un String con la ruta de backup de un álbum de un usuario 
-	 * @param idUsuario Long
-	 * @return String
-	 */
-	public String getBackupUserDirectory(Long idUsuario) {
-		StringBuilder path = new StringBuilder();
-		path.append(backupConfiguration.getAlbum());
-		path.append(File.separatorChar);
-		path.append(idUsuario);
-		
-		return path.toString();
-	}
+//	/**
+//	 * Devuelve un String con la ruta de backup de un álbum de un usuario 
+//	 * @param idUsuario Long
+//	 * @return String
+//	 */
+//	public String getBackupUserDirectory(Long idUsuario) {
+//		StringBuilder path = new StringBuilder();
+//		path.append(backupConfiguration.getAlbum());
+//		path.append(File.separatorChar);
+//		path.append(idUsuario);
+//		
+//		return path.toString();
+//	}
+//
+//	
+//	
+//	
+//	/**
+//	 * Devuelve un String con la ruta de backup de un álbum de un usuario 
+//	 * @param idAlbum Long
+//	 * @param idUsuario Long
+//	 * @return String
+//	 */
+//	public String getBackupAlbumDirectory(Long idAlbum,Long idUsuario) {
+//		StringBuilder path = new StringBuilder();
+//		path.append(backupConfiguration.getAlbum());
+//		path.append(File.separatorChar);
+//		path.append(idUsuario);
+//		
+//		path.append(File.separatorChar);
+//		path.append(idAlbum);
+//		
+//		return path.toString();
+//	}
 
 	
-	
-	
-	/**
-	 * Devuelve un String con la ruta de backup de un álbum de un usuario 
-	 * @param idAlbum Long
-	 * @param idUsuario Long
-	 * @return String
-	 */
-	public String getBackupAlbumDirectory(Long idAlbum,Long idUsuario) {
-		StringBuilder path = new StringBuilder();
-		path.append(backupConfiguration.getAlbum());
-		path.append(File.separatorChar);
-		path.append(idUsuario);
-		
-		path.append(File.separatorChar);
-		path.append(idAlbum);
-		
-		return path.toString();
-	}
-
-	
-	/**
-	 * Devuelve la ruta de backup de una fotografía de un álbum en disco
-	 * @param idAlbum Id del álbum
-	 * @param idUsuario Id del usuario
-	 * @param fileName Nombre de la fotografía
-	 * @return String
-	 */
-	public String getBackupPhoto(Long idAlbum,Long idUsuario,String fileName) {
-		StringBuilder path = new StringBuilder();
-		path.append(getBackupAlbumDirectory(idAlbum, idUsuario));
-		path.append(File.separatorChar);
-		path.append(fileName);
-
-		return path.toString();	
-	}
+//	/**
+//	 * Devuelve la ruta de backup de una fotografía de un álbum en disco
+//	 * @param idAlbum Id del álbum
+//	 * @param idUsuario Id del usuario
+//	 * @param fileName Nombre de la fotografía
+//	 * @return String
+//	 */
+//	public String getBackupPhoto(Long idAlbum,Long idUsuario,String fileName) {
+//		StringBuilder path = new StringBuilder();
+//		path.append(getBackupAlbumDirectory(idAlbum, idUsuario));
+//		path.append(File.separatorChar);
+//		path.append(fileName);
+//
+//		return path.toString();	
+//	}
